# optimized bbox_to_mask by 200x