vectorized the loop over all boxes to optimize the function by 100-200x

https://colab.research.google.com/drive/1IRrpnE9GbCOBRgY3zf44ktELmM_s_YM1?usp=sharing
here's the benchmark + validation script for the optimization 


--- 📊 Benchmark Summary ---
+---------------------------------------------------------------------------------------------------------------+
| Scenario                  | Batch Size   | Mask Size      | Original (ms)   | Optimized (ms)   | Speedup (x)  |
+---------------------------------------------------------------------------------------------------------------+
| Micro Batch (MobileNet)   | 4            | 224x224        | 1.98            | 0.243            | 8.2          |
| Single Image Inference    | 8            | 480x640        | 3.35            | 0.256            | 13.1         |
| Instance Segmentation     | 16           | 256x256        | 4.70            | 0.243            | 19.4         |
| Webcam Detection (HD)     | 64           | 720x1280       | 19.89           | 0.175            | 113.7        |
| Batch Processing (FHD)    | 128          | 1080x1920      | 40.88           | 0.193            | 212.2        |
| Stress Test (Many Boxes)  | 2048         | 512x512        | 650.63          | 0.230            | 2828.9       |
+---------------------------------------------------------------------------------------------------------------+